Changes required to allow DriverView test cases to run with TeamCity

+++ selenium-wd/tests/common/source/TestSetup.java
@@ -58,13 +58,8 @@ import org.openqa.selenium.remote.RemoteWebDriver;
  */
 public class TestSetup {
 
-	private static final String CI_HOST_ADDRESS = "20.20.10.82";
+	private static final String CI_HOST_ADDRESS = "20.20.20.59";
 	private static String testPropFileName;
-	
-	//public static final String SIMULATOR_TEST_REPLAY_FOLDER = "C:\\PicarroAnalyzerSimAutomation";
-	//public static final String SIMULATOR_TEST_REPLAY_DB3_BAT_FILE = "C:\\PicarroAnalyzerSimAutomation\\replay-db3-curl.bat";
-	//public static final String SIMULATOR_TEST_DEFN_FILE = "C:\\PicarroAnalyzerSimAutomation\\replay-db3.defn";
-	//public static final String SIMULATOR_TEST_DB3_FILE = "C:\\PicarroAnalyzerSimAutomation\\Surveyor.db3";
 	
 	public static final String REPLAY_DEFN_CURL_FILE = "replay-defn-curl.bat";
 	public static final String ANALYZER_EXE_PATH = "C:\\PicarroAnalyzer\\Picarro.Surveyor.Analyzer.exe";
@@ -212,21 +207,25 @@ public class TestSetup {
 		String batchFileDirectory = getExecutionPath(getRootPath()) + "lib" + File.separator;
 
 		String command = null;
-		if (!isRunningLocally()) {
-			command = "SetupSimulatorPreReqs.cmd";
-		} else {
+		if (isRunningLocally()) {
 			command = "SetupSimulatorPreReqs-Local.cmd";
+			ProcessUtility.executeProcess(batchFileDirectory + command, /*isShellCommand*/ true, /*waitForExit*/ true);
 		}
-		 
-		ProcessUtility.executeProcess(batchFileDirectory + command, /*isShellCommand*/ true, /*waitForExit*/ true);
 	}
 	
 	public static void startAnalyzer() throws IOException {
-		// Kill any existing instance of Analyzer if running.
-		stopAnalyzer();
-		
-		// Start the Analyzer process.
-		analyzerProcess = ProcessUtility.executeProcess(ANALYZER_EXE_PATH, /*isShellCommand*/ false, /*waitForExit*/ false);	
+		if (isRunningLocally()) {
+			// Kill any existing instance of Analyzer if running.
+			stopAnalyzer();
+			
+			// Start the Analyzer process.
+			analyzerProcess = ProcessUtility.executeProcess(ANALYZER_EXE_PATH, /*isShellCommand*/ false, /*waitForExit*/ false);
+			if (analyzerProcess.isAlive()) {
+				System.out.println("Analyzer EXE started Successfully!");
+			} else {
+				System.out.println("Analyzer EXE did NOT start.");
+			}
+		}
 	}
 
 	public static void replayDB3Script(String defnFileName, String db3FileName) {
@@ -275,8 +274,14 @@ public class TestSetup {
 	}
 
 	public static void stopAnalyzer() {
-		ProcessUtility.killProcess("Picarro.Surveyor.Analyzer.exe", /*killChildProcesses*/ true);
-		ProcessUtility.killProcess("Supervisor.exe", /*killChildProcesses*/ true);
+		try {
+			if (isRunningLocally()) {
+				ProcessUtility.killProcess("Picarro.Surveyor.Analyzer.exe", /*killChildProcesses*/ true);
+				ProcessUtility.killProcess("Supervisor.exe", /*killChildProcesses*/ true);
+			}
+		} catch (UnknownHostException e) {
+			e.printStackTrace();
+		}
 	}
 
 	public void startReplay(String defnFileName) throws InstantiationException, IllegalAccessException, IOException {
